+++ R/CAGEexp.R
@@ -174,7 +174,8 @@ setMethod( "initialize", "CAGEexp"
 #'   CTSStoGenes()                                   |>
 #'   normalizeTagCount()                             |>
 #'   getExpressionProfiles("CTSS")                   |>
-#'   clusterCTSS()                                   |>
+#'   filterLowExpCTSS()                              |>
+#'   distclu()                                       |>
 #'   annotateTagClusters(exampleZv9_annot)           |>
 #'   cumulativeCTSSdistribution("tagClusters")       |>
 #'   quantilePositions("tagClusters")                |>